+++ packages/v1-sdk/src/actions/settle/estimateSettleGasUnits.ts
@@ -5,6 +5,14 @@ export const estimateSettleGasUnits = async (
   peripheryContract: ethers.Contract,
   settlePeripheryParams: SettlePeripheryParams,
 ): Promise<BigNumber> => {
+  const unsignedTx: ethers.PopulatedTransaction =
+    await peripheryContract.populateTransaction.settlePositionAndWithdrawMargin(
+      settlePeripheryParams.marginEngineAddress,
+      settlePeripheryParams.positionOwnerAddress,
+      settlePeripheryParams.tickLower,
+      settlePeripheryParams.tickUpper,
+    );
+
   const estimatedGas: BigNumber = await peripheryContract.estimateGas
     .settlePositionAndWithdrawMargin(
       settlePeripheryParams.marginEngineAddress,
@@ -14,8 +22,9 @@ export const estimateSettleGasUnits = async (
     )
     .catch((error) => {
       console.error(error);
+      console.log(unsignedTx);
       throw new Error(
-        'Error estimating settle position and withdraw margin gas units',
+        'Error estimating settle position and withdraw margin gas units.',
       );
     });