+++ packages/v1-sdk/src/actions/settle/settle.ts
@@ -14,6 +14,7 @@ import { PERIPHERY_ADDRESS_BY_CHAIN_ID } from '../../common/constants';
 import { PositionInfo } from '../../common/api/position/types';
 import { getPositionInfo } from '../../common/api/position/getPositionInfo';
 import { getSentryTracker } from '../../init';
+import { getReadableErrorMessage } from '../../common/errors/errorHandling';
 
 export const settle = async ({
   positionId,
@@ -25,20 +26,18 @@ export const settle = async ({
 
   const positionInfo: PositionInfo = await getPositionInfo(positionId);
 
+  // todo: use decode of id
   const chainId: number = await signer.getChainId();
 
   const peripheryAddress: string = PERIPHERY_ADDRESS_BY_CHAIN_ID[chainId];
 
+  // todo: use decode of id
   const positionOwnerAddress: string = await signer.getAddress();
-
-  const provider: providers.Provider = signer.provider;
 
   const peripheryContract: ethers.Contract = getPeripheryContract(
     peripheryAddress,
-    provider,
+    signer,
   );
-
-  peripheryContract.connect(signer);
 
   const settlePeripheryParams: SettlePeripheryParams = {
     marginEngineAddress: positionInfo.ammMarginEngineAddress,
@@ -46,6 +45,19 @@ export const settle = async ({
     tickLower: positionInfo.positionTickLower,
     tickUpper: positionInfo.positionTickUpper,
   };
+
+  await peripheryContract.callStatic
+    .settlePositionAndWithdrawMargin(
+      settlePeripheryParams.marginEngineAddress,
+      settlePeripheryParams.positionOwnerAddress,
+      settlePeripheryParams.tickLower,
+      settlePeripheryParams.tickUpper,
+    )
+    .catch((error) => {
+      const errorMessage = getReadableErrorMessage(error);
+      console.log(errorMessage);
+      // throw new Error(errorMessage);
+    });
 
   const settlePeripheryTempOverrides: {
     value?: BigNumber;